+++ app/controllers/messages_plivo_settings_controller.rb
@@ -3,7 +3,7 @@ class MessagesPlivoSettingsController < ApplicationController
 
   CHANNEL_KEY = "plivo".freeze
   SETTING_NAMES = %w[
-    messages_plivo_enabled messages_plivo_incoming
+    messages_plivo_enabled messages_plivo_incoming messages_plivo_default_country_only
     messages_plivo_auth_id messages_plivo_auth_token messages_plivo_from
   ].freeze
 end