+++ app/controllers/messages_textanywhere_settings_controller.rb
@@ -3,7 +3,7 @@ class MessagesTextanywhereSettingsController < ApplicationController
 
   CHANNEL_KEY = "textanywhere".freeze
   SETTING_NAMES = %w[
-    messages_textanywhere_enabled messages_textanywhere_incoming
+    messages_textanywhere_enabled messages_textanywhere_incoming messages_textanywhere_default_country_only
     messages_textanywhere_api_key messages_textanywhere_from
   ].freeze
 end